finished till rr interval list
